Add Runtime_events.EV_FUTEX_WAIT and EV_EMPTY_MINOR
EV_FUTEX_WAIT is useful to see in traces when OCaml asked the OS to
suspend a domain. If the domain is waiting for another domain on the
same CPU, this is the point where it is likely to be scheduled.

EV_EMPTY_MINOR shows when a domain is trying to empty its minor heap. It
may be a long time between starting this process and actually performing
a minor GC if, for example, another domain is holding the platform lock.
Without this event, profiling tools tend to under-report the amount of
time spent on GC.